Refactor LoginItem to fix issue with removing login item
For code cleanliness, the choice was originally made to use computed properties to get the login item list and the transitions item within in.
However, these properties returned new instances every time. Apparently some logic in the LSSharedFileListItemRemove function will not work when you pass it an item whose reference is not in the list. 
The item is not in the list since technically it was from a new instance of it.

Fixes an issue where disabling the login item had no effect.