Refactor slightly to enable UI refresh feature
Refactor StatusBarController, TransitionsController, UserData to enable menulet to refresh app start at login when the popover is shown.

Fixes an issue where the toggle would not be updated in the UI when the user removes or adds the app in the Login items
section and then shows the popover again. 

The UI now triggers a refresh when the popover is shown.