CoreBrightnessDiag.StatusInfo decoding changes
On Intel and Apple Silicon the output of corebrightnessdiag seems to be slightly different, with the plist not having a
`Display` key entry sometimes. In this commit, the decoding process has been changed to hopefully more robustly look for just the needed fields (DisplayServicesBrightness & DisplayServicesIsBuiltInDisplay) that seem to always be present.
They are just in different places it seems sometimes.